remove revgeo
Instead of `revgeo::revgeo()`, we will use `tmaptools::rev_geocode_OSM()` in chapter 14. This way we spare one dependency (especially because we would have to use the gitub version of **revgeo** and it does not seem that the package is well maintained at the moment) since we already use **tmaptools** as a dependency.